[REF]anuncioslumicolor: Modifications to the pdf of quotation and transfers
Field delivery_days is added, in the quotation; to have the time, in number of days, for the delivery of the product.
Image field is added, in the budget lines; to add a photo per article per quotation.

The images are shown in the quotation and transfer pdfs.
The delivery_days field is shown in pdf.
Items are listed in quotation and transfer pdfs.
The font size of terms and conditions is reduced (note field).

The td_taxes column is hidden in the pdf.
The customer_lead column is hidden, in the quotation lines.