[REF] anuncioslumicolor: Fonts in Sales Report
Adding the files for a new font called 'Amino' to allow setting it
in the Sales Report and changing the font size to set 8pt as
default and change in specific sections.

Adding a specific class to the header, article and footer in the
external layout that is being used when the record is from a `sale.order`.

The files with the format `.ttf` were giving an error when converting
the HTML to PDF (wkhtmltopdf render), some of the characters
(like C, O, 0, or parentheses) were cut off, so the `.otf` were used.
https://github.com/wkhtmltopdf/wkhtmltopdf/issues/4583